requirements.txt
boto==2.29.1
flask==0.10.1
raven==5.0.0
gunicorn==19.0.0
blinker==1.3
gevent==1.0.1
py_lru_cache==0.1
